+++ dbms/src/Debug/DAGProperties.h
@@ -1,0 +1,34 @@
+// Copyright 2022 PingCAP, Ltd.
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+#pragma once
+
+#include <Core/Defines.h>
+#include <Storages/Transaction/Types.h>
+
+namespace DB
+{
+struct DAGProperties
+{
+    String encode_type;
+    Int64 tz_offset = 0;
+    String tz_name;
+    Int32 collator = 0;
+    bool is_mpp_query = false;
+    bool use_broadcast_join = false;
+    Int32 mpp_partition_num = 1;
+    Timestamp start_ts = DEFAULT_MAX_READ_TSO;
+    Int32 mpp_timeout = 10;
+};
+} // namespace DB
